fix: Flickering images/unstable scrolling
This fixes a bug with the image scanning that caused the CodeMirror
editor to be refreshed every scanning interval (3s by default) if there
were any images scrolled out of view. This had a negative interaction
with the new sync-scrolling which resulting in the editor "jumping"
after scroll events.
After this change, the refresh function will only be called when an
image has changed size AND is visible. In typical usage, the refresh
function will never be called.