fix: Scrolling still stutters
The problem was that I was checking the width/height of the img before it
reloads, and then it checks the clientWidth/clientHeight of the
resulting image. This behaviour was wrong because these fields (while
often the same) are not always going to be the same which results in
unnecessary refresh.

I also removed the width checks because they're not strictly necessary.
The editor only needs to be refreshed on visible height changes, thus
checking im.height is sufficient.