Add list indentation for CodeMirror 6
Looks like this is a bit trickier than cm5, and it could be related to
how Joplin is placing the editor, I'm not sure. However, the tab stop
behaviour seems to have changed with the new editor, so any line that
uses tab indents can't quite get the correct indentation. It will indent
to the level of the tab, but not account for the list marker.

I also needed to update the app min version because versions 3.3 and
prior has some funkiness related to padding-left, likely related to
https://github.com/laurent22/joplin/pull/12746